+++ changelog.md
@@ -12,6 +12,8 @@
   - remove connInterval() is replaced by BLEConnection's getConnInterval()
   - Central.disconnect() is repalced by Bluefruit.disconnect()
   - add conn_handle to Bluefruit.requestPairing(), remove Gap.requestPairing()
+  - Gap.setAddr()/getAddr() are replaced by Bluefruit's setAddr()/getAddr()
+  - Gap.getPeerName() is replaced by Bluefruit.getPeerName()
 - Enhance Gap, add BLEConnection class to mange peer connection
   - Remove Gap.connected(), Gap.getRole() in prefer to BLEConnection API
 - Added setRssiCallback(), monitorRssi(), getRssi(), stopRssi() for tracking rssi of a connection